backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
playwright==1.47.0
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
